+++ L2J_Mobius_1.0_Ertheia/java/com/l2jmobius/gameserver/model/actor/instance/L2RaidBossInstance.java
@@ -102,7 +102,7 @@ public class L2RaidBossInstance extends L2MonsterInstance
 	}
 	
 	@Override
-	public int getVitalityPoints(int level, long exp, boolean isBoss)
+	public int getVitalityPoints(int level, double exp, boolean isBoss)
 	{
 		return -super.getVitalityPoints(level, exp, isBoss);
 	}
