+++ L2J_Mobius_3.0_Helios/java/com/l2jmobius/gameserver/model/actor/instance/L2GrandBossInstance.java
@@ -54,7 +54,7 @@ public final class L2GrandBossInstance extends L2MonsterInstance
 	}
 	
 	@Override
-	public int getVitalityPoints(int level, long exp, boolean isBoss)
+	public int getVitalityPoints(int level, double exp, boolean isBoss)
 	{
 		return -super.getVitalityPoints(level, exp, isBoss);
 	}
